change type variables to double and add filter to 'valueIT'

+++ src/main/java/com/CezaryZal/dto/DimensionDTO.java
@@ -2,9 +2,9 @@ package com.CezaryZal.dto;
 
 public interface DimensionDTO {
 
-int getValueOfDimension();
+double getValueOfDimension();
 
-int getLowerDeviation();
+double getLowerDeviation();
 
-int getUpperDeviation();
+double getUpperDeviation();
 }
